minor fixes: compiles on amd64 (but doesn't run), fixed memory leak with opengl   textures (they were never freed), untested drawCircle in visual.d, simplified   weaponsel.d, sth. in wwp converter I don't remember
svn-rev: 427

+++ src/gui/edit.d
@@ -297,9 +297,9 @@ uint findPrevWord(char[] str, uint pos) {
 }
 
 //little helpers
-bool isSpace(char[] s, uint pos) {
+bool isSpace(char[] s, size_t pos) {
     return isspace(utf.decode(s, pos)) != 0;
 }
-bool isWord(char[] s, uint pos) {
+bool isWord(char[] s, size_t pos) {
     return isUniAlpha(utf.decode(s, pos)) != 0;
 }
